feat: move ThreadWithResult to utils
